feat: Refactor component imports and enhance serialization utilities
