+++ src/core/components/Panel.ts
@@ -1,5 +1,4 @@
-
-import { ControlProps, withControl } from '.';
+import { ControlProps, withControl } from './control';
 import { FunctionComponent, JSX } from '../../jsx';
 
 export interface PanelProps extends ControlProps {
