+++ src/core/index.ts
@@ -1,0 +1,31 @@
+// Components
+export {
+  Fragment,
+  Image,
+  Panel,
+  Text,
+  type FragmentProps,
+  type ImageProps,
+  type PanelProps,
+  type TextProps,
+  type ControlProps,
+} from './components';
+
+// Serialization
+export {
+  reserveBytes,
+  serialize,
+  PROTOCOL_HEADER,
+  VERSION,
+  PAD_CHAR,
+  TYPE_WIDTH,
+  PREFIX_WIDTH,
+  MARKER_WIDTH,
+  FULL_WIDTH,
+  TYPE_PREFIX,
+  FIELD_MARKERS,
+  PROTOCOL_HEADER_LENGTH,
+} from './serializer';
+
+// Rendering
+export { render } from './render';
