+++ src/jsx/index.ts
@@ -1,5 +1,2 @@
 export { renderJSX } from './jsx-runtime';
 export type { FunctionComponent, JSX } from './jsx-runtime';
-
-export { useContext } from './context';
-export type { Context } from './context';
