+++ src/util/index.ts
@@ -1,0 +1,1 @@
+export { Logger } from './Logger';
